// finanzas_frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        AlertaPagoEstadoEnum: "Pendiente" | "Enviada" | "Leida" | "Resuelta";
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}